@@ -2,8 +2,8 @@ package com.setminusx.ramsey.mw.utility;
 
 import com.setminusx.ramsey.mw.entity.Edge;
 
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
