@@ -1,8 +1,8 @@
 package com.setminusx.ramsey.mw.entity;
 
+import jakarta.persistence.*;
 import lombok.Data;
 
-import javax.persistence.*;
 import java.util.Date;
 
 @Data
